requirements.txt
pocketflow>=0.0.1
pydantic>=1.10.0
openai>=1.0.0
PyYAML>=6.0
python-dotenv>=1.0.0
pyvespa>=0.41.0
Pillow>=9.0.0